# schemas/openapi/api.yaml
openapi: 3.0.3
info:
  title: TodoApi
  description: TodoApi
  version: 1.0.0
servers:
  - url: 'http://localhost:8090'
paths:
  /todos:
    get:
      summary: search todos
      operationId: retrieveTodos
      parameters:
        - in: query
          name: filter
          schema:
            type: string
          required: false


      responses:
        200:
          description: Successful Query
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueryPage'


    post:
      summary: Create a todo
      operationId: createTodo
      requestBody:
        description: create a new todo
        content:
          application/json:
            schema:
              title: NewTodoRequest
              type: object
              properties:
                details:
                  type: string
                  description: Textual description of the todo item. unique for incomplete todo
      responses:
        201:
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResource'

  /todos/{todoId}:
    get:
      operationId: retrieveTodo
      summary: retrieve a todo by it's id
      parameters:
        - in: path
          name: todoId
          schema:
            type: string
          required: true

      responses:
        200:
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResource'


  /todos/{todoId}/complete:
    post:
      summary: Complete an existing todo
      operationId: completeTodo
      parameters:
        - in: path
          name: todoId
          schema:
            type: string
          required: true

      responses:
        201:
          description: Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TodoResource'


components:
  schemas:
    QueryPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/TodoResource'

    TodoResource:
      type: object
      properties:
        creator:
          type: string
          format: uri
        details:
          type: string
          description: Textual description of the todo item. unique for incomplete todo
        id:
          type: string
        completed:
          type: boolean